Refactored monolithic FederationManager class into separate chunks for code readbility and maintainability.
